Rebalanced tests
